real_time_weather_monitoring/requirements.txt
kafka-python==2.0.2
requests==2.28.1
psycopg2==2.9.3

